backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0